app/requirements.txt
alembic==1.10.2
anyio==3.6.2
asyncio==3.4.3
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==3.4.8
databases==0.7.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.94.0
fastapi-jwt-auth==0.5.0
fuzzywuzzy==0.18.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
Jinja2==3.1.2
jwt-pydantic==0.0.6
Levenshtein==0.20.9
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.6
PyJWT==1.7.1
python-jose==3.3.0
python-Levenshtein==0.20.9
python-multipart==0.0.6
pytz==2023.3
rapidfuzz==2.15.0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.26.0.post1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.0














